refactor(markdown): name the writers after what they bind, not their role
morphir-langkit-markdown-compiler-scalatags becomes
morphir-langkit-markdown-scalatags, and the kyo-ui writer likewise. The
`compiler` segment restated what every writer in this family is, so it carried
no information a reader did not already have from the coordinate's shape, and
it made a five-segment name where four say the same thing: this binds the
Markdown langkit to a rendering library.

Directories move with the names -- markdown/scalatags and markdown/kyo/ui --
so artifactName still falls out of the path with no override, and the
markdown/compiler namespace marker is gone. Scala packages follow:
morphir.langkit.markdown.{scalatags, kyoui}.

The `markdown` segment stays. These artifacts serve the Markdown langkit
specifically -- the Compiler algebra is Markdown-shaped, with heading,
paragraph and fencedCode -- so a name claiming to be the langkit-wide binding
would promise something the code does not keep.

Nothing published carries the old coordinates yet, so no consumer breaks.
