--- morphir/langkit/markdown/compiler/kyo/ui/test/src/morphir/langkit/markdown/compiler/kyoui/KyoUiCompilerTests.scala
@@ -1,95 +1,0 @@
-package morphir.langkit.markdown.compiler.kyoui
-
-import kyo.*
-import kyo.test.*
-import morphir.langkit.core.Span
-import morphir.langkit.markdown.*
-
-/**
- * These assert on the shape of the `kyo.UI` tree and on the tags that reach the page — never on whole rendered
- * documents byte for byte.
- *
- * That is deliberate. kyo-ui stamps a `data-kyo-path` attribute onto every element it emits, so a byte-exact
- * expectation here would be pinning a reactive framework's bookkeeping rather than our node mapping. The ScalaTags
- * writer carries the byte-exact duty, and the shared algebra is what keeps the two mappings from drifting.
- */
-class KyoUiCompilerTests extends Test[Any]:
-
-  private val span = Span.zero
-
-  private def compile(blocks: Block*): UI =
-    KyoUiCompiler.compile(Document(Chunk.from(blocks), span))
-
-  /** The first emission of a static render: kyo-ui only emits again when a signal changes. */
-  private def render(blocks: Block*): String < Async =
-    UI.runRender(compile(blocks*)).take(1).run.map(_.head)
-
-  private def children(ui: UI): Chunk[UI] = ui match
-    case UI.Ast.Fragment(cs) => cs
-    case other               => Chunk(other)
-
-  "KyoUiCompiler node mapping" - {
-
-    "compiles a heading to the kyo-ui element for its level" in {
-      assert(children(compile(Block.Heading(HeadingLevel.One, "Title", span))).head.isInstanceOf[UI.Ast.H1])
-      assert(children(compile(Block.Heading(HeadingLevel.Two, "Title", span))).head.isInstanceOf[UI.Ast.H2])
-      assert(children(compile(Block.Heading(HeadingLevel.Six, "Title", span))).head.isInstanceOf[UI.Ast.H6])
-    }
-
-    "compiles a paragraph to a P" in
-      assert(children(compile(Block.Paragraph("Body", span))).head.isInstanceOf[UI.Ast.P])
-
-    "compiles a thematic break to an Hr" in
-      assert(children(compile(Block.ThematicBreak(span))).head.isInstanceOf[UI.Ast.Hr])
-
-    "compiles a fenced code block to a Pre wrapping a Code" in {
-      val pre = children(compile(Block.FencedCode(FenceInfo.empty, "x", span))).head
-      assert(pre.isInstanceOf[UI.Ast.Pre])
-      assert(pre.asInstanceOf[UI.Ast.Pre].children.head.isInstanceOf[UI.Ast.Code])
-    }
-
-    "compiles a bullet list to a Ul holding one Li per item" in {
-      val list = children(compile(Block.UnorderedList(Chunk("one", "two"), span))).head
-      assert(list.isInstanceOf[UI.Ast.Ul])
-      val items = list.asInstanceOf[UI.Ast.Ul].children
-      assert(items.size == 2)
-      assert(items.forall(_.isInstanceOf[UI.Ast.Li]))
-    }
-
-    "keeps every top-level block, in order" in {
-      val compiled = children(
-        compile(Block.Heading(HeadingLevel.One, "T", span), Block.Paragraph("B", span), Block.ThematicBreak(span))
-      )
-      assert(compiled.size == 3)
-      assert(compiled(0).isInstanceOf[UI.Ast.H1])
-      assert(compiled(1).isInstanceOf[UI.Ast.P])
-      assert(compiled(2).isInstanceOf[UI.Ast.Hr])
-    }
-  }
-
-  "KyoUiCompiler rendering" - {
-
-    "emits the heading tag and its text" in
-      render(Block.Heading(HeadingLevel.Two, "Title", span)).map { html =>
-        assert(html.startsWith("<h2"))
-        assert(html.contains("Title"))
-        assert(html.endsWith("</h2>"))
-      }
-
-    "escapes text rather than emitting it raw" in
-      render(Block.Paragraph("a < b & c", span)).map { html =>
-        assert(html.contains("&lt;"))
-        assert(html.contains("&amp;"))
-        assert(!html.contains("a < b"))
-      }
-
-    "puts the fence language in a class on the code element" in
-      render(Block.FencedCode(FenceInfo.parse("scala"), "x", span)).map { html =>
-        assert(html.contains("language-scala"))
-        assert(html.contains("<pre"))
-        assert(html.contains("<code"))
-      }
-
-    "renders an empty document without raising" in
-      render().map(html => assert(html.isEmpty))
-  }
